用户列表完成

+++ src/main/java/com/cxb/storehelperserver/repository/RoleRepository.java
@@ -7,7 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
 
 import javax.annotation.Resource;
-import java.util.List;
+import java.util.*;
 
 /**
  * desc: 角色仓库
@@ -20,8 +20,11 @@ public class RoleRepository extends BaseRepository<TRole> {
     @Resource
     private TRoleMapper roleMapper;
 
+    private final String cacheGroupName;
+
     public RoleRepository() {
         init("role::");
+        cacheGroupName = cacheName + "group::";
     }
 
     public TRole find(int id) {
@@ -38,6 +41,20 @@ public class RoleRepository extends BaseRepository<TRole> {
         return tRole;
     }
 
+    public List<TRole> findByGroup(int gid) {
+        List<TRole> roles = List.class.cast(redisTemplate.opsForValue().get(cacheName + cacheGroupName + gid));
+        if (null != roles) {
+            return roles;
+        }
+        TRoleExample example = new TRoleExample();
+        example.or().andGidEqualTo(gid);
+        roles = roleMapper.selectByExample(example);
+        if (null != roles) {
+            redisTemplate.opsForValue().set(cacheName + cacheGroupName + gid, roles);
+        }
+        return roles;
+    }
+
     public List<TRole> all(int gid, String search) {
         TRoleExample example = new TRoleExample();
         if (null == search) {
@@ -51,6 +68,7 @@ public class RoleRepository extends BaseRepository<TRole> {
     public boolean insert(TRole row) {
         if (roleMapper.insert(row) > 0) {
             setCache(row.getId(), row);
+            delCache(cacheName + cacheGroupName + row.getGid());
             return true;
         }
         return false;
@@ -59,12 +77,19 @@ public class RoleRepository extends BaseRepository<TRole> {
     public boolean update(TRole row) {
         if (roleMapper.updateByPrimaryKey(row) > 0) {
             setCache(row.getId(), row);
+            delCache(cacheName + cacheGroupName + row.getGid());
             return true;
         }
         return false;
     }
 
     public boolean delete(int id) {
+        TRole role = find(id);
+        if (null == role) {
+            return false;
+        }
+        delCache(cacheName + cacheGroupName + role.getGid());
+
         delCache(id);
         if (roleMapper.deleteByPrimaryKey(id) <= 0) {
             return false;
